+++ src/parser.rs
@@ -31,6 +31,7 @@ pub(crate) mod go;
 mod helper;
 pub(crate) mod java;
 pub(crate) mod js;
+pub(crate) mod kotlin;
 pub(crate) mod objective_c;
 pub(crate) mod python;
 pub(crate) mod rust;